Reordered download_firmware_validate and restoring the lte_coex_backup register. Realtek does it in this order and based on my testing with RTL8821CU, sometimes the current order can cause issues with firmware validation. Also increased the delay and number of testing attempts for check_hw_ready, once again based on what Realtek does.
References:
https://github.com/morrownr/8821cu-20210916/blob/main/hal/halmac/halmac_88xx/halmac_fw_88xx.c
Line 205 restores the lte_coe_backup. The equivalent of the check in download_firmware_validate happens inside the call to flfw_end_flow_88xx on line 201, the opposite order compared to what rtw88 was doing.
Line 666 sets the check loop, equivalent to rtw88's download_firmware_validate function to 5000 cycles and the delay used is 50 uS on line 678.